Fixed SwiftLint errors and fixed inspector in Sonoma
